# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wedding invitation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RsvpInput {
  name: string;
  attendanceChoice: string;
  attendanceStatus: string;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  guestQuota?: number | null;
  /** @nullable */
  guestCount?: number | null;
}

export interface Guest {
  _id?: string;
  name?: string;
  attendanceStatus?: string;
  attendanceChoice?: string;
  /** @nullable */
  note?: string | null;
  isCheckedIn?: boolean;
  /** @nullable */
  checkedInAt?: string | null;
  /** @nullable */
  checkInMethod?: string | null;
  /** @nullable */
  ticketToken?: string | null;
  /** @nullable */
  guestQuota?: number | null;
  /** @nullable */
  guestCount?: number | null;
  createdAt?: string;
}

export interface RsvpResponse {
  success?: boolean;
  rsvp?: Guest;
}

export interface Wish {
  _id?: string;
  name?: string;
  note?: string;
  createdAt?: string;
}

export interface Photo {
  _id?: string;
  url?: string;
  guestId?: string;
  createdAt?: string;
}

export interface PhotoUploadInput {
  fileBase64?: string;
  guestId?: string;
}

export interface PhotoUploadResponse {
  message?: string;
  photo?: Photo;
}

export interface CheckinInput {
  token: string;
}

export interface CheckinResponse {
  success?: boolean;
  message?: string;
  guest?: Guest;
}

export interface AdminLoginInput {
  username: string;
  password: string;
}

export interface AdminTokenResponse {
  access_token?: string;
  username?: string;
}

export interface GuestListResponse {
  guests?: Guest[];
  total?: number;
  page?: number;
  totalPages?: number;
}

export interface GuestInput {
  name: string;
  attendanceStatus: string;
  attendanceChoice: string;
  /** @nullable */
  note?: string | null;
}

export interface GuestUpdate {
  /** @nullable */
  attendanceStatus?: string | null;
  /** @nullable */
  attendanceChoice?: string | null;
  /** @nullable */
  isCheckedIn?: boolean | null;
  /** @nullable */
  note?: string | null;
}

export type StatsByEvent = {
  gereja?: number;
  resepsi?: number;
  keduanya?: number;
};

export interface Stats {
  total?: number;
  attending?: number;
  notAttending?: number;
  checkedIn?: number;
  /** @nullable */
  totalGuestCount?: number | null;
  byEvent?: StatsByEvent;
}

export interface QRCheckinInput {
  qrData: string;
}

export type GetGuestsParams = {
page?: number;
limit?: number;
search?: string;
};