# lib/api-zod/src/generated/types/checkinResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wedding invitation API
 * OpenAPI spec version: 0.1.0
 */
import type { Guest } from './guest';

export interface CheckinResponse {
  success?: boolean;
  message?: string;
  guest?: Guest;
}