# lib/api-zod/src/generated/types/guestUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wedding invitation API
 * OpenAPI spec version: 0.1.0
 */

export interface GuestUpdate {
  /** @nullable */
  attendanceStatus?: string | null;
  /** @nullable */
  attendanceChoice?: string | null;
  /** @nullable */
  isCheckedIn?: boolean | null;
  /** @nullable */
  note?: string | null;
}