# lib/api-zod/src/generated/types/rsvpInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wedding invitation API
 * OpenAPI spec version: 0.1.0
 */

export interface RsvpInput {
  name: string;
  attendanceChoice: string;
  attendanceStatus: string;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  guestQuota?: number | null;
  /** @nullable */
  guestCount?: number | null;
}