# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wedding invitation API
 * OpenAPI spec version: 0.1.0
 */
import type { StatsByEvent } from './statsByEvent';

export interface Stats {
  total?: number;
  attending?: number;
  notAttending?: number;
  checkedIn?: number;
  /** @nullable */
  totalGuestCount?: number | null;
  byEvent?: StatsByEvent;
}